pull out test suite into own file to use for all tests
